chore: Reuse sha256 digest from spin_loader
