+++ crates/cloud/Cargo.toml
@@ -25,7 +25,6 @@ reqwest = { version = "0.11", features = ["stream"] }
 semver = "1.0"
 serde = {version = "1.0", features = ["derive"]}
 serde_json = "1.0"
-sha2 = "0.9"
 spin-loader = { path = "../loader" }
 spin-publish = { path = "../publish" }
 tempfile = "3.3.0"
