+++ crates/loader/src/lib.rs
@@ -13,6 +13,7 @@
 mod assets;
 pub mod bindle;
 mod common;
+pub mod digest;
 pub mod local;
 
 /// Load a Spin application configuration from a spin.toml manifest file.
@@ -20,8 +21,6 @@ pub use local::from_file;
 
 /// Load a Spin application configuration from Bindle.
 pub use crate::bindle::from_bindle;
-
-pub use crate::assets::file_sha256_digest_string;
 
 /// Maximum number of assets to process in parallel
 pub(crate) const MAX_PARALLEL_ASSET_PROCESSING: usize = 16;
