+++ crates/publish/Cargo.toml
@@ -15,7 +15,6 @@ path-absolutize = "3.0.11"
 reqwest = "0.11"
 semver = "1.0"
 serde = { version = "1.0", features = [ "derive" ] }
-sha2 = "0.10.1"
 spin-loader = { path = "../loader" }
 tempfile = "3.3.0"
 tokio = "1.16.1"
